feat: Update the required endpoints

+++ Backend/TechnicalSupport/Domain/Model/Command/CreateTechnicianCommand.cs
@@ -6,4 +6,4 @@ namespace Backend.TechnicalSupport.Domain.Model.Command;
 /// <param name="Name"></param>
 /// <param name="Status"></param>
 /// <param name="Stars"></param>
-public record CreateTechnicianCommand(string Name, bool Status, int Stars);
+public record CreateTechnicianCommand(string Name, bool Status, double Stars);
