+++ Backend/TechnicalSupport/Domain/Model/Command/UpdateTechnicianCommand.cs
@@ -11,5 +11,5 @@ public record UpdateTechnicianCommand(
     int Id,
     string Name,
     bool Status,
-    int Stars
+    double Stars
     );
