+++ Backend/TechnicalSupport/Interfaces/REST/Resources/CreateTechnicianResource.cs
@@ -1,3 +1,3 @@
 namespace Backend.TechnicalSupport.Interfaces.REST.Resources;
 
-public record CreateTechnicianResource(string Name, bool Status, int Stars);
+public record CreateTechnicianResource(string Name, bool Status, double Stars);
